Revert "[refactor] Move components from ./src/components/common to ./src/components."
This reverts commit d0bc256fa820d856fdd50a0cfa1bfd619bc66c34.
