+++ src/components/BackgroundScreen/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import BackgroundImage from '../BackgroundImage';
+import BackgroundImage from '../common/BackgroundImage';
 
 const BackgroundScreen = () => {
     return(
