+++ src/components/BackgroundScreen/styles.js
