+++ src/components/common/NavigatorComponent.js
